bag fix, refactiring, error message, validation fields

+++ src/pages/main/work/credit/pages/registration/PreApproval.vue
@@ -99,7 +99,7 @@
                 color="green"
                 v-close-popup
                 :disable="disableBtn"
-                @click="successCredit(false)"
+                @click="successCredit"
               />
               <q-btn
                 label="Отменить"
@@ -125,6 +125,9 @@ import CommonUtils from "@/shared/utils/CommonUtils";
 // import LoaderFullScreen from "@/components/LoaderFullScreen";
 
 export default {
+  props: {
+    confirm: Boolean
+  },
   data() {
     return {
       failureCreditReason: false,
@@ -144,9 +147,6 @@ export default {
     disableBtn() {
       return this.$store.getters["credits/credits"].disableBtn;
     },
-    confirm() {
-      return this.$store.getters["credits/credits"].confirm;
-    },
     preApprovalData() {
       return this.$store.getters["credits/credits"].preApprovalData;
     },
@@ -158,9 +158,9 @@ export default {
     }
   },
   methods: {
-    async successCredit(val) {
+    async successCredit() {
       console.log(this.$store);
-      this.$store.commit("credits/toggleConfirm", val);
+      //this.confirm = false
       this.$emit("toggleLoaderForm", true);
       console.log(JSON.stringify(this.credits.confirmCreditData, null, 2));
       try {
@@ -171,47 +171,48 @@ export default {
 
         console.log("response", response);
 
-        const data = response.nextTask.input.find(
-          i => i.label === "application"
-        ).data;
-        const dictionaries = response.nextTask.input.find(
-          i => i.label === "inputDictionaries"
-        ).data;
+        if (response) {
+          const data = response.nextTask.input.find(
+            i => i.label === "application"
+          ).data;
+          const dictionaries = response.nextTask.input.find(
+            i => i.label === "inputDictionaries"
+          ).data;
 
-        console.log("dic", JSON.stringify(dictionaries, null, 2));
+          console.log("dic", JSON.stringify(dictionaries, null, 2));
 
-        this.$store.commit("profile/setPreapprovData", data);
-        this.$store.commit("profile/setDictionaries", dictionaries);
+          this.$store.commit("profile/setPreapprovData", data);
+          this.$store.commit("profile/setDictionaries", dictionaries);
 
-        sessionStorage.setItem("preapprovData", JSON.stringify(data));
-        sessionStorage.setItem("dictionaries", JSON.stringify(dictionaries));
-        
-        this.$router.push("profile");
+          sessionStorage.setItem("preapprovData", JSON.stringify(data));
+          sessionStorage.setItem("dictionaries", JSON.stringify(dictionaries));
+          
+          this.$router.push("profile");
+          this.$emit("toggleLoaderForm", false);
+        }
+      } catch (error) {
+        this.$store.commit("credits/setMessage", CommonUtils.filterServerError(error));
         this.$emit("toggleLoaderForm", false);
-        
-      } catch (error) {}
+      }
     },
 
     async failureCredit() {
       this.$refs.toggle.validate();
       if (this.$refs.toggle.hasError) {
         this.formHasError = true;
-        //this.$store.commit("toggleConfirm", true);
       } else {
         this.$emit("toggleLoaderFullScreen", true);
         this.$store.commit("credits/toggleDisableInput", false);
-        this.$store.commit("credits/toggleConfirm", false);
-
         this.credits.confirmCreditData.output[0].data = false;
         this.credits.confirmCreditData.output[1].data = this.selection;
 
         console.log(JSON.stringify(this.credits.confirmCreditData, null, 2));
         try {
-          const res = await this.$store.dispatch(
+          const response = await this.$store.dispatch(
             "credits/confirmationCredit",
             this.credits.confirmCreditData
           );
-          console.log("res", res);
+          console.log("res", response);
           
           // if (res.requestedTask.state === "completed") {
           //   this.$store.commit("credits/setMessage", "Credit failure");
@@ -222,10 +223,16 @@ export default {
           //   throw "Task do not completed";
           // }
 
-          this.$store.commit("credits/setMessage", "Credit failure");
-          sessionStorage.clear();
-          this.$router.push("/work/credit");
-        } catch (error) {}
+          if (response) {
+            this.$store.commit("credits/setMessage", "Credit failure");
+            sessionStorage.clear();
+            this.$router.push("/work/credit");
+          }
+          
+        } catch (error) {
+          this.$emit("toggleLoaderFullScreen", false);
+          this.$store.commit("credits/setMessage", CommonUtils.filterServerError(error));
+        }
       }
     },
 
@@ -246,17 +253,22 @@ export default {
             "credits/getFile",
             this.fileData
           );
-
-          this.fileData.idFile = file.id
+          
+          if (file) {
+            this.fileData.idFile = file.id // для кеширования id
+          }
         }
 
         console.log('file', file)
 
-        printJS(file.url);
-        window.URL.revokeObjectURL(file.url);
+        if (file) {
+          printJS(file.url);
+          window.URL.revokeObjectURL(file.url);
+        }
 
         this.loading = false
       } catch(error) {
+        this.$store.commit("credits/setMessage", CommonUtils.filterServerError(error));
         this.loading = false
       }
     }
